+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 * Scope the results view and export to only return active students in the current college ([#885](https://gitlab.com/yale-sdmp/vesta/issues/885)).
+* Ensure that suites are properly treated as available after archiving old groups ([#895](https://gitlab.com/yale-sdmp/vesta/issues/895)).
 
 ## v2.1.0 - 2019-02-04
 ### Notes for upgrade: